fix: code Optimization
