+++ examples/cos-batch-uploader/components/FileUpload/DropZone.tsx
@@ -4,6 +4,7 @@ import { useCallback, useState } from 'react'
 import { Upload, FileImage, FileVideo, AlertCircle } from 'lucide-react'
 import { clsx } from 'clsx'
 import { Button } from '../UI/Button'
+import { formatFileSize } from '../../lib/utils'
 
 interface DropZoneProps {
   onFilesSelected: (files: File[]) => void
@@ -68,13 +69,7 @@ export function DropZone({
     e.target.value = ''
   }, [])
 
-  const formatFileSize = (bytes: number) => {
-    if (bytes === 0) return '0 Bytes'
-    const k = 1024
-    const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB']
-    const i = Math.min(Math.floor(Math.log(bytes) / Math.log(k)), sizes.length - 1)
-    return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + ' ' + sizes[i]
-  }
+
 
   // Check storage capacity
   const checkStorageCapacity = async (files: File[]): Promise<StorageCheckResult> => {
