+++ examples/cos-batch-uploader/components/History/ExportDialog.tsx
@@ -5,6 +5,7 @@ import { Download, X } from 'lucide-react'
 import { Modal } from '../UI/Modal'
 import { Button } from '../UI/Button'
 import { HistoryItem } from '../../types/upload'
+import { formatFileSize } from '../../lib/utils'
 
 interface ExportDialogProps {
   isOpen: boolean
@@ -79,13 +80,7 @@ export function ExportDialog({ isOpen, onClose, items, selectedCount = 0 }: Expo
     document.body.removeChild(link)
   }
 
-  const formatFileSize = (bytes: number) => {
-    if (bytes === 0) return '0 Bytes'
-    const k = 1024
-    const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB']
-    const i = Math.min(Math.floor(Math.log(bytes) / Math.log(k)), sizes.length - 1)
-    return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + ' ' + sizes[i]
-  }
+
 
   const getExportTitle = () => {
     return 'Export File Records'
