+++ examples/cos-batch-uploader/components/History/HistoryStats.tsx
@@ -1,22 +1,14 @@
 'use client'
 
 import { HistoryItem } from '../../types/upload'
+import { formatFileSize } from '../../lib/utils'
 
 interface HistoryStatsProps {
   items: HistoryItem[]
 }
 
 export function HistoryStats({ items }: HistoryStatsProps) {
-  const formatFileSize = (bytes: number) => {
-    // Validate input
-    if (!bytes || bytes === 0 || !isFinite(bytes)) return '0 Bytes'
-    if (bytes < 0) return '0 Bytes'
-    
-    const k = 1024
-    const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB']
-    const i = Math.min(Math.floor(Math.log(bytes) / Math.log(k)), sizes.length - 1)
-    return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + ' ' + sizes[i]
-  }
+
 
   // Calculate total size with validation
   const totalSize = items.reduce((sum, item) => {
